+++ app/imports/api/objects.js
@@ -4,11 +4,8 @@ import { Mongo } from 'meteor/mongo';
 import { Locations } from '/imports/api/locations.js'; // , geoJSONPointSchema
 import { getUserDescription } from '/imports/api/users.js';
 import { Integrations } from '/imports/api/integrations.js';
-import { CoinSchema } from '/imports/api/bikecoin.js';
-
-if(Meteor.isServer) {
-  import BikeCoin from '/server/api/BikeCoin.js';
-}
+import { CoinSchema } from '/imports/api/settings.js';
+import BikeCoin from '/imports/api/BikeCoin.js';
 
 export const Objects = new Mongo.Collection('objects');
