+++ app/imports/api/settings.js
@@ -1,7 +1,6 @@
 import { Meteor } from 'meteor/meteor'
 import { Mongo } from 'meteor/mongo';
 import { getUserDescription } from '/imports/api/users.js';
-import { CoinSchema } from '/imports/api/bikecoin.js';
 
 export const Settings = new Mongo.Collection('settings');
 
@@ -45,6 +44,19 @@ if (Meteor.isServer) {
 		return Settings.findOne({profileName: profileName});
 	}
 }
+
+export const CoinSchema = new SimpleSchema({
+  'address': {
+    type: String,
+    label: "Address",
+    defaultValue: ''
+  },
+  'privatekey': {
+    type: String,
+    label: "Private Key",
+    defaultValue: ''
+  }
+});
 
 export const MapboxSchema = new SimpleSchema({
   'style': {
